+++ HISTORY.rst
@@ -1,11 +1,19 @@
 =======
 History
 =======
+2.2.0 (2020-09-24)
+------------------
+* [MOD] add support for prepending and appending mappings on input dataframe (Mapper)
+* [MOD] add support for custom spark sql functions in mapper without injecting methods
+* [MOD] add support for "on"/"off" and "enabled"/"disabled" in extended_string_to_boolean custom mapper transformations
+* [ADD] new custom mapper transformations:
+    - extended_string_to_date
+    - extended_string_unix_timestamp_ms_to_date
 
 2.1.1 (2020-09-04)
 ------------------
-* [ADD] `drop_rows_with_empty_array` flag to allow keeping rows with empty array after explosion
-* [ADD] additional test-cases for extended_string mappings (non string inputs)
+* [MOD] `drop_rows_with_empty_array` flag to allow keeping rows with empty array after explosion
+* [MOD] additional test-cases for extended_string mappings (non string inputs)
 * [FIX] remove STDERR logging, don't touch root logging level anymore (needs to be done outside spooq to see some lower log levels)
 * [ADD] new custom mapper transformations:
     - extended_string_unix_timestamp_ms_to_timestamp
@@ -13,8 +21,8 @@ History
 2.1.0 (2020-08-17)
 ------------------
 * [ADD] Python 3 support
-* [ADD] `ignore_missing_columns` flag to fail on missing input columns with Mapper transformer (https://github.com/Breaka84/Spooq/pull/6)
-* [ADD] timestamp support for threshold cleaner
+* [MOD] `ignore_missing_columns` flag to fail on missing input columns with Mapper transformer (https://github.com/Breaka84/Spooq/pull/6)
+* [MOD] timestamp support for threshold cleaner
 * [ADD] new custom mapper transformations:
     - meters_to_cm
     - unix_timestamp_ms_to_spark_timestamp